refactoring file structure.
