+++ lib/segments/fans/aggregateSegmentResults.ts
