+++ lib/segments/fans/createFanSegments.ts
