+++ lib/segments/fans/getLLMResponse.ts
