+++ lib/segments/fans/parseAndValidateResponse.ts
