+++ lib/segments/fans/processBatchesInParallel.ts
