+++ lib/segments/fans/processSegmentResults.ts
