+++ lib/segments/fans/validateComments.ts
