+++ lib/segments/fans/validateFanSocialIds.ts
